+++ src/Tab/stories/index.jsx
@@ -1,0 +1,6 @@
+import { storiesOf } from '@storybook/react'
+
+import overview from './overview'
+
+storiesOf('Tabs', module)
+  .add('Overview', overview)